refactor: reorganize the helpers
