+++ components/helpers/index.js
@@ -3,9 +3,3 @@ function newline2br(str = '') {
 }
 
 export { newline2br }
-
-export {
-  setupIntersectionObserver,
-  cleanupIntersectionObserver,
-  isIntersectionObserver,
-} from './intersection-observer.js'
